+++ CHANGELOG.md
@@ -1,3 +1,36 @@
+# [1.2.0](https://github.com/Ex-Populus/ethers-gcp-kms-signer/compare/1.1.2...1.2.0) (2025-03-11)
+
+
+### Bug Fixes
+
+* update import and configuration handling ([79bcffe](https://github.com/Ex-Populus/ethers-gcp-kms-signer/commit/79bcffe439738ef5aaf5d76ce8c14c5f8076f822))
+
+
+### Features
+
+* enhance module compatibility and testing ([dc87a22](https://github.com/Ex-Populus/ethers-gcp-kms-signer/commit/dc87a220dbdb52fed3e3ec6d0766ae12facd456f))
+
+## [1.1.2](https://github.com/Ex-Populus/ethers-gcp-kms-signer/compare/1.1.1...1.1.2) (2025-03-11)
+
+
+### Bug Fixes
+
+* update tsup configuration ([d1cb690](https://github.com/Ex-Populus/ethers-gcp-kms-signer/commit/d1cb6900abba53193939cd77ed38af9fff238fcc))
+
+## [1.1.1](https://github.com/Ex-Populus/ethers-gcp-kms-signer/compare/1.1.0...1.1.1) (2025-03-11)
+
+
+### Bug Fixes
+
+* update package versions and imports ([d47a000](https://github.com/Ex-Populus/ethers-gcp-kms-signer/commit/d47a0006120ccf6b13e182e0801c376e5bdd4c20))
+
+# [1.1.0](https://github.com/Ex-Populus/ethers-gcp-kms-signer/compare/1.0.12...1.1.0) (2025-03-11)
+
+
+### Features
+
+* integrate elliptic library and implement key encoding without using the external library ([4d39f94](https://github.com/Ex-Populus/ethers-gcp-kms-signer/commit/4d39f94692ed4359071f3b582dce1ba06abd5387))
+
 ## [1.0.12](https://github.com/Ex-Populus/ethers-gcp-kms-signer/compare/1.0.11...1.0.12) (2025-03-11)